feat(lib): added AbstractMemoryRepository
