+++ taskforce/apps/tasks/src/app/category/repository/category-memory.repository.ts
@@ -1,28 +1,10 @@
-import { CRUDRepository } from '@taskforce/core';
 import { Category } from '@taskforce/shared-types';
-import * as crypto from 'crypto';
 import { Injectable } from '@nestjs/common';
 import { CategoryEntity } from '../category.entity';
+import { AbstractMemoryRepository } from '@taskforce/core';
 
 @Injectable()
-export class CategoryMemoryRepository implements CRUDRepository<CategoryEntity, string, Category> {
-  private repository: {[key: string]: Category} = {};
-
-  public async create(item: CategoryEntity): Promise<Category> {
-    const entry = { ...item.toObject(), _id: crypto.randomUUID()};
-    this.repository[entry._id] = entry;
-
-    return {...entry};
-  }
-
-  public async findById(id: string): Promise<Category> {
-    if (this.repository[id]) {
-      return {...this.repository[id]};
-    }
-
-    return null;
-  }
-
+export class CategoryMemoryRepository extends AbstractMemoryRepository<CategoryEntity, Category> {
   public async findByName(name: string): Promise<Category | null> {
     const existingCategory = Object.values(this.repository)
       .find((category) => category.name === name);
@@ -36,14 +18,5 @@ export class CategoryMemoryRepository implements CRUDRepository<CategoryEntity, 
 
   public async findAll(): Promise<Category[]> {
     return Object.values(this.repository);
-  }
-
-  public async destroy(id: string): Promise<void> {
-    delete this.repository[id];
-  }
-
-  public async update(id: string, item: CategoryEntity): Promise<Category> {
-    this.repository[id] = {...item.toObject(), _id: id};
-    return this.findById(id);
   }
 }
