+++ taskforce/apps/users/src/app/task-user/repository/task-user-memory.repository.ts
@@ -1,28 +1,10 @@
-import { CRUDRepository } from '@taskforce/core';
 import { User } from '@taskforce/shared-types';
 import { TaskUserEntity } from '../task-user.entity';
-import * as crypto from 'crypto';
 import { Injectable } from '@nestjs/common';
+import { AbstractMemoryRepository } from '@taskforce/core';
 
 @Injectable()
-export class TaskUserMemoryRepository implements CRUDRepository<TaskUserEntity, string, User> {
-  private repository: {[key: string]: User} = {};
-
-  public async create(item: TaskUserEntity): Promise<User> {
-    const entry = { ...item.toObject(), _id: crypto.randomUUID()};
-    this.repository[entry._id] = entry;
-
-    return {...entry};
-  }
-
-  public async findById(id: string): Promise<User> {
-    if (this.repository[id]) {
-      return {...this.repository[id]};
-    }
-
-    return null;
-  }
-
+export class TaskUserMemoryRepository extends AbstractMemoryRepository<TaskUserEntity, User> {
   public async findByEmail(email: string): Promise<User | null> {
     const existUser = Object.values(this.repository)
       .find((userItem) => userItem.email === email);
@@ -32,14 +14,5 @@ export class TaskUserMemoryRepository implements CRUDRepository<TaskUserEntity, 
     }
 
     return { ...existUser};
-  }
-
-  public async destroy(id: string): Promise<void> {
-    delete this.repository[id];
-  }
-
-  public async update(id: string, item: TaskUserEntity): Promise<User> {
-    this.repository[id] = {...item.toObject(), _id: id};
-    return this.findById(id);
   }
 }
